habittracker/requirements.txt
altgraph==0.10.2
asn1crypto==0.24.0
bdist-mpkg==0.5.0
bonjour-py==0.3
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
click==6.7
cryptography==2.7
Django==1.11.29
django-tastypie==0.14.3
djangorestframework==3.9.4
enum34==1.1.6
Flask==1.1.1
future==0.17.1
futures==3.3.0
google-assistant-grpc==0.2.1
google-assistant-sdk==0.5.1
google-auth==1.6.3
google-auth-oauthlib==0.4.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
idna==2.8
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
macholib==1.5.1
MarkupSafe==1.1.1
matplotlib==1.3.1
modulegraph==0.10.4
monotonic==1.5
numpy==1.8.0rc1
oauthlib==3.1.0
pathlib2==2.3.4
protobuf==3.9.1
py2app==0.7.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==19.0.0
pyparsing==2.0.1
python-dateutil==1.5
python-mimeparse==1.6.0
pytz==2013.7
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scandir==1.10.0
scipy==0.13.0b1
six==1.12.0
sounddevice==0.3.13
tenacity==4.12.0
urllib3==1.25.3
Werkzeug==0.15.6
xattr==0.6.4
